# swagger.yaml
openapi: 3.0.0
info:
  title: Music Service API
  description: API для музыкального сервиса
  version: 1.0.0
servers:
  - url: /api/v1
    description: Основной API сервер
paths:
...
  /albums:
    get:
      summary: Получить список всех альбомов
      operationId: listAllAlbums
      tags:
        - albums
      responses:
        '200':
          description: Список всех альбомов
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Album'
        '500':
          description: Ошибка сервера
    
    post:
      summary: Создать новый альбом
      operationId: createAlbum
      security:
        - BearerAuth: []
      tags:
        - albums
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  example: "My New Album"
                artist:
                  type: string
                  example: "Artist Name"
                release_date:
                  type: string
                  format: date-time
                cover_url:
                  type: string
                  example: "https://example.com/album-cover.jpg"
              required:
                - title
                - artist
      responses:
        '201':
          description: Альбом успешно создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Album'
        '400':
          description: Некорректные данные запроса
        '403':
          description: Недостаточно прав (требуются права администратора)
        '500':
          description: Ошибка сервера
  /albums/{id}:
    get:
      summary: Получить детали альбома
      operationId: getAlbumDetails
      tags:
        - albums
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Детали альбома с треками
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlbumWithTracks'
        '404':
          description: Альбом не найден
        '500':
          description: Ошибка сервера
    
    put:
      summary: Обновить информацию об альбоме
      operationId: updateAlbumInfo
      security:
        - BearerAuth: []
      tags:
        - albums
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  example: "Updated Album Title"
                artist:
                  type: string
                  example: "Updated Artist Name"
                release_date:
                  type: string
                  format: date-time
                cover_url:
                  type: string
                  example: "https://example.com/new-cover.jpg"
              required:
                - title
                - artist
      responses:
        '200':
          description: Информация об альбоме обновлена
        '400':
          description: Некорректные данные запроса
        '403':
          description: Недостаточно прав (требуются права администратора)
        '404':
          description: Альбом не найден
        '500':
          description: Ошибка сервера
    
    delete:
      summary: Удалить альбом (вместе со всеми треками)
      operationId: deleteAlbum
      security:
        - BearerAuth: []
      tags:
        - albums
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Альбом и все его треки успешно удалены
        '403':
          description: Недостаточно прав (требуются права администратора)
        '404':
          description: Альбом не найден
        '500':
          description: Ошибка сервера
  /albums/{id}/tracks:
    post:
      summary: Добавить трек в альбом
      operationId: addTrackToAlbum
      security:
        - BearerAuth: []
      tags:
        - albums
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                track_id:
                  type: string
                  format: uuid
              required:
                - track_id
      responses:
        '200':
          description: Трек добавлен в альбом
        '400':
          description: Некорректные данные запроса
        '403':
          description: Недостаточно прав (требуются права администратора)
        '404':
          description: Альбом или трек не найден
        '500':
          description: Ошибка сервера
  /albums/{id}/tracks/{track_id}:
    delete:
      summary: Удалить трек из альбома
      operationId: removeTrackFromAlbum
      security:
        - BearerAuth: []
      tags:
        - albums
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: track_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Трек удален из альбома
        '403':
          description: Недостаточно прав (требуются права администратора)
        '404':
          description: Альбом или трек не найден
        '500':
          description: Ошибка сервера
...
  /genres:
    get:
      summary: Получить список всех жанров
      operationId: listAllGenres
      tags:
        - genres
      responses:
        '200':
          description: Список жанров
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Genre'
        '500':
          description: Ошибка сервера
    
    post:
      summary: Создать новый жанр
      operationId: createGenre
      security:
        - BearerAuth: []
      tags:
        - genres
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: "Rock"
              required:
                - name
      responses:
        '201':
          description: Жанр успешно создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Genre'
        '400':
          description: Некорректные данные запроса
        '403':
          description: Недостаточно прав (требуются права администратора)
        '500':
          description: Ошибка сервера
  /genres/tracks/{id}:
    get:
      summary: Получить жанры для трека
      operationId: getGenresByTrack
      tags:
        - genres
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Список жанров для трека
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Genre'
        '404':
          description: Трек не найден
        '500':
          description: Ошибка сервера
  /genres/tracks/{id}/genres:
    post:
      summary: Назначить жанр треку
      operationId: assignGenreToTrack
      security:
        - BearerAuth: []
      tags:
        - genres
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                genre_id:
                  type: string
                  format: uuid
              required:
                - genre_id
      responses:
        '200':
          description: Жанр назначен треку
        '400':
          description: Некорректные данные запроса
        '403':
          description: Недостаточно прав (требуются права администратора)
        '404':
          description: Трек или жанр не найден
        '500':
          description: Ошибка сервера
  /genres/tracks/{trackId}/genres/{genreId}:
    delete:
      summary: Удалить жанр у трека
      operationId: removeGenreFromTrack
      security:
        - BearerAuth: []
      tags:
        - genres
      parameters:
        - name: trackId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: genreId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Жанр удален у трека
        '403':
          description: Недостаточно прав (требуются права администратора)
        '404':
          description: Трек или жанр не найден
        '500':
          description: Ошибка сервера
...
components:
  schemas:
...
    Track:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        duration:
          type: integer
          description: Продолжительность трека в секундах
        file_path:
          type: string
          description: Относительный путь к MP3 файлу
        album_id:
          type: string
          format: uuid
          description: ID альбома, к которому принадлежит трек (обязательно)
        artist_name:
          type: string
        cover_url:
          type: string
          description: URL обложки трека
        added_date:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        play_count:
          type: integer
          description: Количество прослушиваний
        genres:
          type: array
          items:
            $ref: '#/components/schemas/Genre'
          description: Список жанров трека
      required:
        - id
        - title
        - duration
        - file_path
        - artist_name
        - added_date
        - updated_at
...
    Album:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        artist:
          type: string
        release_date:
          type: string
          format: date-time
        cover_url:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - title
        - artist
        - release_date
    AlbumWithTracks:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        artist:
          type: string
        release_date:
          type: string
          format: date-time
        cover_url:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        tracks:
          type: array
          items:
            $ref: '#/components/schemas/Track'
      required:
        - id
        - title
        - artist
        - release_date
        - tracks
...
    Genre:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
      required:
        - id
        - name
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT 